LPS-83223 update test names and usage
