+++ modules/apps/bookmarks/bookmarks-test/src/testIntegration/java/com/liferay/bookmarks/internal/exportimport/test/BookmarksExportImportTest.java
@@ -56,7 +56,7 @@ public class BookmarksExportImportTest extends BasePortletExportImportTestCase {
 
 	@Override
 	public String getPortletId() {
-		return BookmarksPortletKeys.BOOKMARKS;
+		return BookmarksPortletKeys.BOOKMARKS_ADMIN;
 	}
 
 	@Before
